requirements.txt
pyspark
psycopg2
psycopg2-binary==2.9.9
apache-airflow-providers-apache-spark
python-dotenv==1.0.1